Atualizar o Q08.py

+++ Q08.py
@@ -1,5 +1,5 @@
-n1 = int(input("Diga-me um número:"))
-n2 = int(input("Diga-me outro número:"))
+n1 = float(input("Diga-me um número:"))
+n2 = float(input("Diga-me outro número:"))
 print(f"A soma de {n1} e {n2}:", (n1 + n2))
 print(f"A subtração de {n1} e {n2}:" , (n1 - n2))
 print(f"A multiplicação entre {n1} e {n2}:" , (n1 * n2))
